Only build windows launcher on windows...
git-svn-id: http://svn.osafoundation.org/chandler/trunk@847 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/__hardhat__.py
@@ -44,20 +44,21 @@ def build(buildenv):
 
 
     # Build the windows launcher program
-    version = buildenv['version']
+    if buildenv['os'] == 'win':
+        version = buildenv['version']
 
-    try:
-        os.remove('output.txt')
-    except:
-        pass
+        try:
+            os.remove('output.txt')
+        except:
+            pass
 
-    hardhatlib.executeCommand(buildenv, info['name'],
-                              [ buildenv['compiler'],
-                                'distrib/win/launcher/launcher.sln',
-                                '/build', version.capitalize(),
-                                '/out', 'output.txt' ],
-                              'Building launcher ' + version,
-                              0, 'output.txt')
+        hardhatlib.executeCommand(buildenv, info['name'],
+                                  [ buildenv['compiler'],
+                                    'distrib/win/launcher/launcher.sln',
+                                    '/build', version.capitalize(),
+                                    '/out', 'output.txt' ],
+                                  'Building launcher ' + version,
+                                  0, 'output.txt')
 
 
     # Build UUID Extension and install it
